# Item attribute system redesign, data-driven multipliers, and 32-bit damage pipeline
- Replaced packed bitmask item attributes with individual fields (prefix, secondary, enchant, custom-made)
- Added migration script for converting existing account databases
- Fixed migration to store raw nibble values, preventing a double-scaling bug

- Made attribute multipliers data-driven from the database instead of hardcoded
- Multipliers can be tuned and hot-reloaded via "reload attributes" without recompiling
- All prefix and secondary effects now use the DB-driven multiplier lookup

- Widened damage pipeline from 8/16-bit to 32-bit to prevent truncation of large values
- Sharp and Ancient prefix bonuses now add deterministic flat damage instead of random dice range

+++ Sources/Dependencies/Shared/Packet/PacketEvent.h
@@ -90,7 +90,7 @@ namespace net {
 		PacketHeader header;
 		std::uint16_t object_id;
 		std::uint8_t dir;
-		std::uint8_t v1;
+		std::int32_t v1;
 		std::uint8_t v2;
 		std::int16_t x;
 		std::int16_t y;
@@ -100,7 +100,7 @@ namespace net {
 		PacketHeader header;
 		std::uint16_t object_id;
 		std::uint8_t dir;
-		std::int16_t v1;
+		std::int32_t v1;
 		std::uint8_t v2;
 	};
